Handle user authentication and authorization
API clients should be able to generate authentication tokens to be used in
subsequent calls to authenticated endpoints.

Since this is the first feature that interacts with the User resource we
had to implement the following changes:
  * Create User resource schema and database table
  * Handle password encryption

We also implemented an Auth Plug that can be used to ensure the call is
authorized before getting to the controller action. This plug can be
used as a step in the endpoint pipeline.